fix code warning

+++ core/connection/ble/src/softbus_ble_connection.c
@@ -1023,7 +1023,7 @@ static void BleOnDataReceived(bool isBleConn, int32_t halConnId, uint32_t len, c
             RecvConnectedComd(targetNode->connId, (const cJSON*)data);
             cJSON_Delete(data);
         } else {
-            if (head->len + sizeof(ConnPktHead) > (int32_t)len) {
+            if (head->len + sizeof(ConnPktHead) > len) {
                 SoftBusLog(SOFTBUS_LOG_CONN, SOFTBUS_LOG_ERROR, "BLEINFOPRTINT: recv a big data:%d, not support",
                     head->len + sizeof(ConnPktHead));
             }
